FinTrack/Backend_new/requirements/requirements.txt
# Core Django
Django>=5.0,<6.0
djangorestframework>=3.15,<4.0

# Authentication
djangorestframework-simplejwt>=5.3,<6.0
django-cors-headers>=4.3,<5.0
dj-rest-auth>=6.0,<7.0

# Database
psycopg2-binary>=2.9,<3.0
dj-database-url>=2.1,<3.0

# Environment variables
python-decouple>=3.8,<4.0

# API Documentation
drf-yasg>=1.21,<2.0

# Filtering and pagination
django-filter>=24.0,<25.0

# Data export
openpyxl>=3.1,<4.0
reportlab>=4.1,<5.0

# Date utilities
python-dateutil>=2.8,<3.0

# Validation
phonenumbers>=8.13,<9.0

# Production
gunicorn>=22.0,<23.0
whitenoise>=6.6,<7.0